Add files via upload
Personal Diary Management System:

In this console application, you can add, view, edit and delete records. Records can be added with many information such as duration of task, name, address, time and date. File handling has been effectively used to keep the records.
The basic user-defined functions used in this project are listed below:
•	int password() – contains/manages/handles password protection
•	void addrecord() – to add new diary record
•	void viewrecord() – to view added record in list
•	void editrecord() – to modify and update an added record
•	void editpassword() – to modify/change a password
•	void deleterecord() – to delete or remove a record permanently from system file

In this project, you need to enter a password to view, edit or delete record. Password is not required to view the main menu or to add record. What’s little different in this project is that you can edit the password within the application i.e., after executing or upon running the project.
Your password is currently ‘Enter’ for this mini project. Don’t type e-n-t-e-r. Press the Enter key